refactor: extract hardcoded content into centralized data config
  - Move all TypeScript interfaces to src/data/content.types.ts
  - Re-export types from content.ts via export type *
  - Update Facilities.tsx to use siteContent with dynamic icon resolution
  - Update Programs.tsx to use siteContent for program data
  - Update Pricing.tsx to use siteContent with currency and CTA fields
  - Update Testimonials.tsx to use siteContent.testimonials.reviews
  - Update Contact.tsx to use siteContent with getIcon and mapEmbedUrl
  - Update Footer.tsx to use siteContent for links, social, and brand data
  - Update layout.tsx to source metadata from siteContent.meta
  - Make Hero logo and heading sizes viewport-height responsive via clamp()
  - Update Google Maps embed URL to correct AlphaTime location

  Resolves: #1
