+++ src/data/content.ts
@@ -8,232 +8,9 @@
 // Components render a styled placeholder when image is null.
 // ============================================================================
 
-// ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
+import type { SiteContent } from "./content.types";
 
-export interface SiteMetadata {
-  title: string;
-  description: string;
-  keywords: string[];
-}
-
-export interface BrandInfo {
-  name: string;
-  subtitle: string;
-  tagline: string;
-  motto: string;
-  logo: string | null; // path to logo image, null = use icon fallback
-  watermark: string; // short text used as watermark in placeholders
-}
-
-export interface SocialLink {
-  platform: "instagram" | "facebook" | "tiktok" | "google_maps";
-  label: string;
-  href: string;
-}
-
-export interface NavLink {
-  name: string;
-  href: string;
-}
-
-export interface NavContent {
-  links: NavLink[];
-  cta: { label: string; href: string };
-}
-
-export interface HeroStat {
-  value: string;
-  label: string;
-}
-
-export interface HeroContent {
-  tagline: string;
-  headingLine1: string;
-  headingLine2: string;
-  description: string;
-  ctaPrimary: { label: string; href: string };
-  ctaSecondary: { label: string; href: string };
-  stats: HeroStat[];
-  backgroundImage: string | null;
-}
-
-export interface AboutValue {
-  iconName: string; // lucide icon name
-  title: string;
-  description: string;
-}
-
-export interface AboutContent {
-  label: string;
-  headingMain: string;
-  headingAccent: string;
-  description: string;
-  image: string | null;
-  badge: { line1: string; line2: string };
-  values: AboutValue[];
-}
-
-export interface FacilityItem {
-  iconName: string;
-  title: string;
-  description: string;
-  image: string | null;
-}
-
-export interface FacilitiesContent {
-  label: string;
-  headingMain: string;
-  headingAccent: string;
-  description: string;
-  items: FacilityItem[];
-}
-
-export interface ProgramItem {
-  number: string;
-  title: string;
-  subtitle: string;
-  description: string;
-  tags: string[];
-  image: string | null;
-}
-
-export interface ProgramsContent {
-  label: string;
-  headingMain: string;
-  headingAccent: string;
-  description: string;
-  items: ProgramItem[];
-}
-
-export interface TrainerItem {
-  name: string;
-  role: string;
-  specialties: string[];
-  experience: string;
-  image: string | null;
-  socialHref: string | null;
-}
-
-export interface TrainersContent {
-  label: string;
-  headingMain: string;
-  headingAccent: string;
-  description: string;
-  members: TrainerItem[];
-}
-
-export interface PricingPlan {
-  name: string;
-  price: string;
-  currency: string;
-  period: string;
-  description: string;
-  features: string[];
-  popular: boolean;
-  ctaLabel: string;
-  ctaHref: string;
-}
-
-export interface PricingContent {
-  label: string;
-  headingMain: string;
-  headingAccent: string;
-  description: string;
-  note: string;
-  popularBadge: string;
-  plans: PricingPlan[];
-}
-
-export interface GalleryItem {
-  title: string;
-  subtitle: string;
-  span: string; // tailwind grid span classes
-  image: string | null;
-}
-
-export interface GalleryContent {
-  label: string;
-  headingMain: string;
-  headingAccent: string;
-  description: string;
-  items: GalleryItem[];
-}
-
-export interface TestimonialItem {
-  name: string;
-  role: string;
-  text: string;
-  rating: number;
-  avatar: string | null;
-}
-
-export interface TestimonialsContent {
-  label: string;
-  headingMain: string;
-  headingAccent: string;
-  description: string;
-  reviews: TestimonialItem[];
-}
-
-export interface ContactInfoBlock {
-  iconName: string;
-  title: string;
-  details: string[];
-}
-
-export interface ContactFormField {
-  name: string;
-  label: string;
-  type: string;
-  placeholder: string;
-  colSpan?: number; // 1 or 2 (full width)
-}
-
-export interface ContactContent {
-  label: string;
-  headingMain: string;
-  headingAccent: string;
-  description: string;
-  info: ContactInfoBlock[];
-  mapEmbedUrl: string;
-  formFields: ContactFormField[];
-  interests: string[];
-  submitLabel: string;
-}
-
-export interface FooterLinkGroup {
-  title: string;
-  links: { name: string; href: string }[];
-}
-
-export interface FooterContent {
-  ctaHeadingLine1: string;
-  ctaHeadingLine2: string;
-  ctaLabel: string;
-  ctaHref: string;
-  description: string;
-  linkGroups: FooterLinkGroup[];
-  copyright: string;
-}
-
-export interface SiteContent {
-  meta: SiteMetadata;
-  brand: BrandInfo;
-  social: SocialLink[];
-  nav: NavContent;
-  hero: HeroContent;
-  about: AboutContent;
-  facilities: FacilitiesContent;
-  programs: ProgramsContent;
-  trainers: TrainersContent;
-  pricing: PricingContent;
-  gallery: GalleryContent;
-  testimonials: TestimonialsContent;
-  contact: ContactContent;
-  footer: FooterContent;
-}
+export type * from "./content.types";
 
 // ---------------------------------------------------------------------------
 // Content
@@ -681,7 +458,7 @@ export const siteContent: SiteContent = {
       },
     ],
     mapEmbedUrl:
-      "https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3402.5!2d74.35!3d31.52!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x0%3A0x0!2zMzHCsDMxJzEyLjAiTiA3NMKwMjEnMDAuMCJF!5e0!3m2!1sen!2spk!4v1",
+      "https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3402.0911500821526!2d74.40456929999999!3d31.494178299999994!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x3919054b14f0c6bb%3A0x527c71b017d1f37c!2sAlphatime%20Fitness%20Club%20(Gym%20for%20Ladies%20%26%20Gents)!5e0!3m2!1sen!2s!4v1774658676110!5m2!1sen!2s",
     formFields: [
       { name: "name", label: "Full Name", type: "text", placeholder: "Your name", colSpan: 2 },
       { name: "email", label: "Email", type: "email", placeholder: "your@email.com", colSpan: 1 },
